Changed bitset sizes from short to int.

+++ src/af/af.c
@@ -57,7 +57,7 @@ int free_argumentation_framework(AF* af) {
 
 int free_projected_argumentation_framework(PAF* paf) {
 	free(paf->index_mapping);
-	int freed_bytes = paf->af->size * sizeof(unsigned short);
+	int freed_bytes = paf->af->size * sizeof(unsigned int);
 	freed_bytes += free_argumentation_framework(paf->af);
 	free(paf);
 	return freed_bytes + sizeof(PAF);
@@ -102,20 +102,20 @@ PAF* project_argumentation_framework(AF *af, BitSet *mask) {
 	PAF *paf = calloc(1, sizeof(PAF));
 	assert(paf != NULL);
 
-    unsigned short size = count_bits(mask);
+    unsigned int size = count_bits(mask);
 	assert(size > 0);
 
-    paf->index_mapping = calloc(size, sizeof(unsigned short));
-    unsigned short j = 0;
-    for (unsigned short i = 0; i < af->size; ++i) {
+    paf->index_mapping = calloc(size, sizeof(unsigned int));
+    unsigned int j = 0;
+    for (unsigned int i = 0; i < af->size; ++i) {
         if (TEST_BIT(mask, i)) {
             paf->index_mapping[j++] = i;
         }
     }
 
     paf->af = create_argumentation_framework(size);
-    for (unsigned short i = 0; i < size; ++i) {
-        for (unsigned short j = 0; j < size; ++j) {
+    for (unsigned int i = 0; i < size; ++i) {
+        for (unsigned int j = 0; j < size; ++j) {
             if (TEST_BIT(af->graph[paf->index_mapping[i]], paf->index_mapping[j])) {
 				SET_BIT(paf->af->graph[i], j);
             }
@@ -126,9 +126,9 @@ PAF* project_argumentation_framework(AF *af, BitSet *mask) {
 }
 
 
-BitSet* project_back(BitSet* bs, PAF* paf, unsigned short base_size) {
+BitSet* project_back(BitSet* bs, PAF* paf, unsigned int base_size) {
 	BitSet* res = create_bitset(base_size);
-	for (unsigned short i = 0; i < paf->af->size; ++i) {
+	for (unsigned int i = 0; i < paf->af->size; ++i) {
 		if (TEST_BIT(bs, i)) {
 			SET_BIT(res, paf->index_mapping[i]);
 		}
